CurrentMatchコンポーネントでsetMatchWinnerとresetMatchWinnerをuseMatchWinnerフックに移動し、MatchContextを整理
